Fixed package names, added RenderManager
Fixed package names, and then added a RenderManager object.

RenderManager is an object that extends Thread, and it's purpose is to
handle creating the array of triangles that the paint method draws. It
does this constantly, and because it's a thread, frees up the paint
method from having to do these steps each call.